Move rich text helper out into own file
