+++ seeds/tables/utils/to-rich-text.js
@@ -1,0 +1,25 @@
+const toRichText = (...strings) => {
+  return {
+    object: 'value',
+    document: {
+      object: 'document',
+      data: {},
+      nodes: strings.map(text => {
+        return {
+          object: 'block',
+          type: 'paragraph',
+          data: {},
+          nodes: [
+            {
+              object: 'text',
+              text,
+              marks: []
+            }
+          ]
+        };
+      })
+    }
+  };
+};
+
+module.exports = toRichText;
